pyproject.toml
[project]
name = "rag-service"
version = "0.1.0"
description = ""
authors = [
    {name = "exl",email = "levzavadskiy999@gmail.com"}
]
requires-python = ">=3.11,<4"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "sentence-transformers (>=5.3.0,<6.0.0)",
    "chromadb (>=1.5.5,<2.0.0)",
    "gigachat (>=0.2.0,<0.3.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
